sentence_classification_experiment/requirements.txt
bitermplus==0.7.0
jieba==0.42.1
lightgbm==4.3.0
numpy==1.23.1
pandas==2.0.3
parsel==1.9.1
Requests==2.32.3
scikit_learn==1.3.2
tensorflow_gpu==2.9.0
transformers==4.37.2
urllib3==1.26.6
xgboost==2.0.3
